break the circular dependency between reduce and elementwise
