+++ paddle/pten/kernels/gpu/scale_kernel.cu
@@ -16,8 +16,8 @@ limitations under the License. */
 
 #include "paddle/pten/backends/gpu/gpu_context.h"
 #include "paddle/pten/core/kernel_registry.h"
+#include "paddle/pten/kernels/funcs/elementwise_base.h"
 // See Note [ Why still include the fluid headers? ]
-#include "paddle/fluid/operators/elementwise/elementwise_op_impl.cu.h"
 #include "paddle/fluid/platform/float16.h"
 
 namespace pten {
@@ -55,7 +55,7 @@ void ScaleKernel(const ContextT& dev_ctx,
   inputs.emplace_back(&x);
   outputs.emplace_back(out);
   out->mutable_data<T>();
-  LaunchSameDimsElementwiseCudaKernel<ElementwiseType::kUnary, T, T>(
+  funcs::LaunchSameDimsElementwiseCudaKernel<ElementwiseType::kUnary, T, T>(
       dev_ctx,
       inputs,
       &outputs,
